+++ frontend/src/components/layout/AppLayout.vue
@@ -11,17 +11,19 @@ import StatusBar from './StatusBar.vue'
       <Sidebar />
       <main class="min-w-0 flex-1 overflow-hidden">
         <RouterView v-slot="{ Component }">
-          <Transition
-            mode="out-in"
-            enter-active-class="transition duration-150 ease-out"
-            enter-from-class="opacity-0"
-            enter-to-class="opacity-100"
-            leave-active-class="transition duration-100 ease-in"
-            leave-from-class="opacity-100"
-            leave-to-class="opacity-0"
-          >
-            <component :is="Component" :key="$route.path" />
-          </Transition>
+          <KeepAlive>
+            <Transition
+              mode="out-in"
+              enter-active-class="transition duration-150 ease-out"
+              enter-from-class="opacity-0"
+              enter-to-class="opacity-100"
+              leave-active-class="transition duration-100 ease-in"
+              leave-from-class="opacity-100"
+              leave-to-class="opacity-0"
+            >
+              <component :is="Component" :key="$route.path" />
+            </Transition>
+          </KeepAlive>
         </RouterView>
       </main>
     </div>